`total_chapters` to Long

+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/Track.kt
@@ -18,7 +18,7 @@ interface Track : Serializable {
 
     var last_chapter_read: Double
 
-    var total_chapters: Int
+    var total_chapters: Long
 
     var score: Double
