+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/Anilist.kt
@@ -153,7 +153,7 @@ class Anilist(id: Long) : BaseTracker(id, "AniList"), DeletableTracker {
 
         if (track.status != COMPLETED) {
             if (didReadChapter) {
-                if (track.last_chapter_read.toInt() == track.total_chapters && track.total_chapters > 0) {
+                if (track.last_chapter_read.toLong() == track.total_chapters && track.total_chapters > 0) {
                     track.status = COMPLETED
                     track.finished_reading_date = System.currentTimeMillis()
                 } else if (track.status != REREADING) {
