+++ app/src/main/java/eu/kanade/tachiyomi/data/track/kavita/KavitaApi.kt
@@ -93,7 +93,7 @@ class KavitaApi(private val client: OkHttpClient, interceptor: KavitaInterceptor
      * Ignores volumes.
      * Volumes consisting of 1 file treated as chapter
      */
-    private fun getTotalChapters(url: String): Int {
+    private fun getTotalChapters(url: String): Long {
         val requestUrl = getApiVolumesUrl(url)
         try {
             val listVolumeDto = with(json) {
@@ -101,13 +101,13 @@ class KavitaApi(private val client: OkHttpClient, interceptor: KavitaInterceptor
                     .execute()
                     .parseAs<List<VolumeDto>>()
             }
-            var volumeNumber = 0
-            var maxChapterNumber = 0
+            var volumeNumber = 0L
+            var maxChapterNumber = 0L
             for (volume in listVolumeDto) {
                 if (volume.chapters.maxOf { it.number!!.toFloat() } == 0f) {
                     volumeNumber++
                 } else if (maxChapterNumber < volume.chapters.maxOf { it.number!!.toFloat() }) {
-                    maxChapterNumber = volume.chapters.maxOf { it.number!!.toFloat().toInt() }
+                    maxChapterNumber = volume.chapters.maxOf { it.number!!.toFloat().toLong() }
                 }
             }
