+++ app/src/main/java/eu/kanade/tachiyomi/data/track/model/TrackSearch.kt
@@ -18,7 +18,7 @@ class TrackSearch : Track {
 
     override var last_chapter_read: Double = 0.0
 
-    override var total_chapters: Int = 0
+    override var total_chapters: Long = 0
 
     override var score: Double = -1.0
